requirements.txt
django
sklearn
numpy
pandas
djangorestframework